Cleanup gradle files
