+++ buildSrc/src/main/kotlin/ksp-module.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    java
+}
+
+sourceSets {
+    main {
+        java {
+            srcDir(file("$buildDir/generated/ksp/main/kotlin/"))
+        }
+    }
+
+    test {
+        java {
+            srcDir(file("$buildDir/generated/ksp/test/kotlin/"))
+        }
+    }
+}
