+++ buildSrc/src/main/kotlin/published-module.gradle.kts
@@ -1,0 +1,45 @@
+import org.gradle.api.publish.maven.MavenPublication
+
+plugins {
+    `maven-publish`
+    signing
+}
+
+val sourceJar: Task by tasks.getting
+val javadocJar: Task by tasks.getting
+
+publishing {
+    repositories {
+        maven {
+            name = "KotDis"
+
+            url = if (project.version.toString().contains("SNAPSHOT")) {
+                uri("https://maven.kotlindiscord.com/repository/maven-snapshots/")
+            } else {
+                uri("https://maven.kotlindiscord.com/repository/maven-releases/")
+            }
+
+            credentials {
+                username = project.findProperty("kotdis.user") as String? ?: System.getenv("KOTLIN_DISCORD_USER")
+                password = project.findProperty("kotdis.password") as String?
+                    ?: System.getenv("KOTLIN_DISCORD_PASSWORD")
+            }
+
+            version = project.version
+        }
+    }
+
+    publications {
+        create<MavenPublication>("maven") {
+            from(components.getByName("java"))
+
+            artifact(sourceJar)
+            artifact(javadocJar)
+        }
+    }
+}
+
+signing {
+    useGpgCmd()
+    sign(publishing.publications["maven"])
+}
