+++ extra-modules/extra-mappings/build.gradle.kts
@@ -1,12 +1,8 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
 plugins {
-    `maven-publish`
-    signing
-
-    id("io.gitlab.arturbosch.detekt")
-
-    kotlin("jvm")
+    `kordex-module`
+    `published-module`
+    `ksp-module`
+    `dokka-module`
 }
 
 repositories {
@@ -59,79 +55,8 @@ dependencies {
 
 group = "com.kotlindiscord.kord.extensions"
 
-val sourceJar = task("sourceJar", Jar::class) {
-    dependsOn(tasks["classes"])
-    archiveClassifier.set("sources")
-    from(sourceSets.main.get().allSource)
-}
-
-detekt {
-    buildUponDefaultConfig = true
-    config = rootProject.files("detekt.yml")
-
-    autoCorrect = true
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_9
-    targetCompatibility = JavaVersion.VERSION_1_9
-}
-
-tasks.withType<KotlinCompile>().configureEach {
-    kotlinOptions {
-        jvmTarget = "9"
-    }
-}
-
-sourceSets {
-    main {
-        java {
-            srcDir(file("$buildDir/generated/ksp/main/kotlin/"))
-        }
-    }
-
-    test {
-        java {
-            srcDir(file("$buildDir/generated/ksp/test/kotlin/"))
-        }
-    }
-}
-
-tasks.build {
-    this.finalizedBy(sourceJar)
-}
-
-publishing {
-    repositories {
-        maven {
-            name = "KotDis"
-
-            url = if (project.version.toString().contains("SNAPSHOT")) {
-                uri("https://maven.kotlindiscord.com/repository/maven-snapshots/")
-            } else {
-                uri("https://maven.kotlindiscord.com/repository/maven-releases/")
-            }
-
-            credentials {
-                username = project.findProperty("kotdis.user") as String? ?: System.getenv("KOTLIN_DISCORD_USER")
-                password = project.findProperty("kotdis.password") as String?
-                    ?: System.getenv("KOTLIN_DISCORD_PASSWORD")
-            }
-
-            version = project.version
-        }
-    }
-
-    publications {
-        create<MavenPublication>("maven") {
-            from(components.getByName("java"))
-
-            artifact(sourceJar)
-        }
-    }
-}
-
-signing {
-    useGpgCmd()
-    sign(publishing.publications["maven"])
+kordex {
+    jvmTarget.set("9")
+    javaVersion.set(JavaVersion.VERSION_1_9)
+    useExplicitApiMode.set(false)
 }
