+++ gradle.properties
@@ -4,3 +4,6 @@ kotlin.incremental = true
 ksp.incremental = false
 
 projectVersion = 1.5.0-SNAPSHOT
+
+#dokka will run out of memory with the default meta space
+org.gradle.jvmargs=-XX:MaxMetaspaceSize=1024m
