+++ modules/java-time/build.gradle.kts
@@ -1,16 +1,10 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-import java.io.ByteArrayOutputStream
-import java.net.URL
-
 plugins {
-    `maven-publish`
-    signing
+    `kordex-module`
+    `published-module`
+    `dokka-module`
+    `ksp-module`
 
-    kotlin("jvm")
     kotlin("plugin.serialization")
-
-    id("io.gitlab.arturbosch.detekt")
-    id("org.jetbrains.dokka")
 }
 
 dependencies {
@@ -24,149 +18,12 @@ dependencies {
     testImplementation(libs.logback)
 }
 
-val sourceJar = task("sourceJar", Jar::class) {
-    dependsOn(tasks["classes"])
-    archiveClassifier.set("sources")
-    from(sourceSets.main.get().allSource)
+
+kordex {
+    jvmTarget.set("9")
+    javaVersion.set(JavaVersion.VERSION_1_9)
 }
 
-val javadocJar = task("javadocJar", Jar::class) {
-    dependsOn("dokkaJavadoc")
-    archiveClassifier.set("javadoc")
-    from(tasks.javadoc)
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_9
-    targetCompatibility = JavaVersion.VERSION_1_9
-}
-
-kotlin {
-    explicitApi()
-}
-
-tasks.withType<KotlinCompile>().configureEach {
-    kotlinOptions {
-        jvmTarget = "9"
-    }
-}
-
-sourceSets {
-    main {
-        java {
-            srcDir(file("$buildDir/generated/ksp/main/kotlin/"))
-        }
-    }
-
-    test {
-        java {
-            srcDir(file("$buildDir/generated/ksp/test/kotlin/"))
-        }
-    }
-}
-
-detekt {
-    buildUponDefaultConfig = true
-    config = files("../../detekt.yml")
-
-    autoCorrect = true
-}
-
-publishing {
-    repositories {
-        maven {
-            name = "KotDis"
-
-            url = if (project.version.toString().contains("SNAPSHOT")) {
-                uri("https://maven.kotlindiscord.com/repository/maven-snapshots/")
-            } else {
-                uri("https://maven.kotlindiscord.com/repository/maven-releases/")
-            }
-
-            credentials {
-                username = project.findProperty("kotdis.user") as String? ?: System.getenv("KOTLIN_DISCORD_USER")
-                password = project.findProperty("kotdis.password") as String?
-                    ?: System.getenv("KOTLIN_DISCORD_PASSWORD")
-            }
-
-            version = project.version
-        }
-    }
-
-    publications {
-        create<MavenPublication>("maven") {
-            from(components.getByName("java"))
-
-            artifact(sourceJar)
-            artifact(javadocJar)
-        }
-    }
-}
-
-signing {
-    useGpgCmd()
-    sign(publishing.publications["maven"])
-}
-
-fun runCommand(command: String): String {
-    val output = ByteArrayOutputStream()
-
-    project.exec {
-        commandLine(command.split(" "))
-        standardOutput = output
-    }
-
-    return output.toString().trim()
-}
-
-fun getCurrentGitBranch(): String {  // https://gist.github.com/lordcodes/15b2a4aecbeff7c3238a70bfd20f0931
-    var gitBranch = "Unknown branch"
-
-    try {
-        gitBranch = runCommand("git rev-parse --abbrev-ref HEAD")
-    } catch (t: Throwable) {
-        println(t)
-    }
-
-    return gitBranch
-}
-
-tasks.dokkaHtml.configure {
-    moduleName.set("Kord Extensions: Java Time")
-
-    dokkaSourceSets {
-        configureEach {
-            includeNonPublic.set(false)
-            skipDeprecated.set(false)
-
-            displayName.set("Kord Extensions: Java Time")
-            includes.from("packages.md")
-            jdkVersion.set(8)
-
-            sourceLink {
-                localDirectory.set(file("${project.projectDir}/src/main/kotlin"))
-
-                remoteUrl.set(
-                    URL(
-                        "https://github.com/Kotlin-Discord/kord-extensions/" +
-                            "tree/${getCurrentGitBranch()}/java-time/src/main/kotlin"
-                    )
-                )
-
-                remoteLineSuffix.set("#L")
-            }
-
-            externalDocumentationLink {
-                url.set(URL("http://kordlib.github.io/kord/common/common/"))
-            }
-
-            externalDocumentationLink {
-                url.set(URL("http://kordlib.github.io/kord/core/core/"))
-            }
-        }
-    }
-}
-
-tasks.build {
-    this.finalizedBy(sourceJar, javadocJar)
+dokkaModule {
+   moduleName.set("Kord Extensions: Java Time")
 }
